+++ web-ui/src/components/common/BlobDetailsDialog/BlobDetailsDialog.tsx
@@ -1,11 +1,9 @@
-import { BlobIcon } from '../../icons/BlobIcon';
 import { BlobsApi, BlobStatsDto } from '../../../../generated';
-import { DeadBlobIcon } from '../../icons/DeadBlobIcon';
-import { BlobBlinkIcon } from '../../icons/BlobBlinkIcon';
-import { JSX, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { BlobDetailsDialogUi } from './BlobDetailsDialogUi';
 import defaultConfig from '../../../default-config';
 import { useMutation } from '@tanstack/react-query';
+import { BlobAnimated } from '../blob-visuals/BlobAnimated';
 
 interface BlobDetailsDialogProps {
   open: boolean;
@@ -16,7 +14,6 @@ interface BlobDetailsDialogProps {
 
 export const BlobDetailsDialog = ({ open, onClose, cachedBlob, blobId }: BlobDetailsDialogProps) => {
   const [blob, setBlob] = useState<BlobStatsDto>();
-  const [blobIcon, setBlobIcon] = useState<JSX.Element>(<BlobIcon size={180} color={blob?.color ?? ''} />);
 
   const blobsApi: BlobsApi = new BlobsApi(defaultConfig);
   const { mutate: fetchBlobDetails } = useMutation({
@@ -24,7 +21,6 @@ export const BlobDetailsDialog = ({ open, onClose, cachedBlob, blobId }: BlobDet
     onSuccess: (response) => {
       if (response) {
         setBlob(response);
-        setBlobIcon(<BlobIcon size={180} color={response.color} />);
       }
     },
   });
@@ -32,44 +28,19 @@ export const BlobDetailsDialog = ({ open, onClose, cachedBlob, blobId }: BlobDet
   useEffect(() => {
     if (cachedBlob) {
       setBlob(cachedBlob);
-      setBlobIcon(<BlobIcon size={180} color={cachedBlob.color} />);
     } else if (blobId) {
       fetchBlobDetails();
     }
   }, [cachedBlob]);
 
-  useEffect(() => {
-    if (!blob) return;
-    if (blob.isDead) {
-      setBlobIcon(<DeadBlobIcon size={180} color={blob.color} />);
-    } else {
-      setBlobIcon(<BlobIcon size={180} color={blob.color} />);
-      const interval = setInterval(() => {
-        setBlobIcon(<BlobIcon size={180} color={blob.color} />);
-        setTimeout(() => {
-          setBlobIcon(<BlobBlinkIcon size={180} color={blob.color} />);
-        }, 4000);
-        setTimeout(() => {
-          setBlobIcon(<BlobIcon size={180} color={blob.color} />);
-        }, 4250);
-        setTimeout(() => {
-          setBlobIcon(<BlobBlinkIcon size={180} color={blob.color} />);
-        }, 8250);
-        setTimeout(() => {
-          setBlobIcon(<BlobIcon size={180} color={blob.color} />);
-        }, 8500);
-        setTimeout(() => {
-          setBlobIcon(<BlobBlinkIcon size={180} color={blob.color} />);
-        }, 8750);
-        setTimeout(() => {
-          setBlobIcon(<BlobIcon size={180} color={blob.color} />);
-        }, 9000);
-      }, 9000);
-      return () => clearInterval(interval);
-    }
-  }, [blob]);
-
   if (!blob) return null;
 
-  return <BlobDetailsDialogUi open={open} onClose={onClose} blob={blob!} blobIcon={blobIcon} />;
+  return (
+    <BlobDetailsDialogUi
+      open={open}
+      onClose={onClose}
+      blob={blob!}
+      blobIcon={<BlobAnimated blob={blob} size={180} />}
+    />
+  );
 };