+++ web-ui/src/components/common/blob-visuals/blob-parts/Crown.tsx
@@ -1,0 +1,16 @@
+export const Crown = () => {
+  return (
+    <path
+      opacity="1"
+      fill="#ffdf00"
+      stroke="#000000"
+      strokeWidth="1"
+      strokeLinecap="butt"
+      strokeLinejoin="miter"
+      strokeOpacity="1"
+      fillOpacity="1"
+      d="M 35,8 H 95 V -15 L 80,-1 65,-15 50,-1 35,-15 Z"
+      id="path858"
+    />
+  );
+};