+++ web-ui/src/components/common/blob-visuals/blob-parts/eyes/Blink.tsx
@@ -1,0 +1,46 @@
+export const Blink = () => {
+  return (
+    <>
+      <path
+        fill="none"
+        stroke="#000000"
+        strokeWidth="1px"
+        strokeLinecap="butt"
+        strokeLinejoin="miter"
+        strokeOpacity="1"
+        d="m 41,25.5 9,10"
+        id="path992"
+      />
+      <path
+        fill="none"
+        stroke="#000000"
+        strokeWidth="1px"
+        strokeLinecap="butt"
+        strokeLinejoin="miter"
+        strokeOpacity="1"
+        d="m 41,45 9,-10"
+        id="path992-1"
+      />
+      <path
+        fill="none"
+        stroke="#000000"
+        strokeWidth="1px"
+        strokeLinecap="butt"
+        strokeLinejoin="miter"
+        strokeOpacity="1"
+        d="m 89,25.5 -9,10"
+        id="path992-6"
+      />
+      <path
+        fill="none"
+        stroke="#000000"
+        strokeWidth="1px"
+        strokeLinecap="butt"
+        strokeLinejoin="miter"
+        strokeOpacity="1"
+        d="m 80,35 9,10"
+        id="path992-6-1"
+      />
+    </>
+  );
+};