+++ web-ui/src/components/common/blob-visuals/blob-parts/eyes/IdleEyes.tsx
@@ -1,0 +1,26 @@
+export const IdleEyes = () => {
+  return (
+    <>
+      <ellipse
+        fill="#ffffff"
+        stroke="#000000"
+        strokeWidth="1"
+        id="path2446"
+        cx="45"
+        cy="35"
+        rx="4.4549451"
+        ry="9.4549446"
+      />
+      <ellipse
+        fill="#ffffff"
+        stroke="#000000"
+        strokeWidth="1"
+        id="path2446-7"
+        cx="85"
+        cy="35"
+        rx="4.4549451"
+        ry="9.4549446"
+      />
+    </>
+  );
+};