+++ web-ui/src/components/common/blob-visuals/blob-parts/eyes/TerminatedEyes.tsx
@@ -1,0 +1,46 @@
+export const TerminatedEyes = () => {
+  return (
+    <>
+      <path
+        fill="none"
+        stroke="#000000"
+        strokeWidth="1px"
+        strokeLinecap="butt"
+        strokeLinejoin="miter"
+        strokeOpacity="1"
+        d="M 38,28 52,42"
+        id="path1039"
+      />
+      <path
+        fill="none"
+        stroke="#000000"
+        strokeWidth="1px"
+        strokeLinecap="butt"
+        strokeLinejoin="miter"
+        strokeOpacity="1"
+        d="M 38,42 52,28"
+        id="path1039-7"
+      />
+      <path
+        fill="none"
+        stroke="#000000"
+        strokeWidth="1px"
+        strokeLinecap="butt"
+        strokeLinejoin="miter"
+        strokeOpacity="1"
+        d="M 78,28 92,42"
+        id="path1039-6"
+      />
+      <path
+        fill="none"
+        stroke="#000000"
+        strokeWidth="1px"
+        strokeLinecap="butt"
+        strokeLinejoin="miter"
+        strokeOpacity="1"
+        d="M 78,42 92,28"
+        id="path1039-6-2"
+      />
+    </>
+  );
+};